scripts/pyproject.toml
[project]
name = "oj-scripts"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "aiohttp-socks>=0.9.0",
    "beautifulsoup4>=4.13.3",
    "colorlog>=6.9.0",
    "curl-cffi>=0.7.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pytz>=2025.2",
    "requests>=2.32.3",
    "sqlite-vec>=0.1.6",
    "tenacity>=8.2.0",
    "tomli>=2.0.1; python_version < '3.11'",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
